+++ py2app/recipes/__init__.py
@@ -30,3 +30,4 @@ from . import uuid  # noqa: F401
 from . import virtualenv  # noqa: F401
 from . import wx  # noqa: F401
 from . import xml  # noqa: F401
+from . import pycryptodome  # noqa: F401